requirements.txt
Django>=3.2.3,<3.3
psycopg2==2.9.5
psycopg2-binary==2.9.6
openpyxl==3.1.5
pandas==2.2.2
requests==2.32.3
#django-tailwind
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.13.0,<4
